+++ test/integration-tests/cdk-upgrade-from-0.9.4.test.ts
@@ -1,8 +1,7 @@
 import { PerformTasksCommand, ValidateTasksCommand, RemoveCommand } from '~commands/index';
 import { IIntegrationTestContext, baseBeforeAll, baseAfterAll } from './base-integration-test';
 import { ChildProcessUtility } from '~util/child-process-util';
-import { GetObjectOutput } from 'aws-sdk/clients/s3';
-import { ExecOptions } from 'child_process';
+import { GetObjectCommand, GetObjectCommandOutput } from '@aws-sdk/client-s3';
 
 const basePathForScenario = './test/integration-tests/resources/scenario-cdk-no-region/';
 
@@ -10,9 +9,9 @@ describe('when calling org-formation perform tasks', () => {
     let context: IIntegrationTestContext;
 
     let spawnProcessAfterDeploy: jest.MockContext<any, any>;
-    let stateAfterDeploy: GetObjectOutput;
+    let stateAfterDeploy: GetObjectCommandOutput;
     let spawnProcessMock: jest.SpyInstance;
-    let stateAfterCleanup: GetObjectOutput;
+    let stateAfterCleanup: GetObjectCommandOutput;
     let spawnProcessAfterCleanup: jest.MockContext<any, any>;
 
     beforeAll(async () => {
@@ -27,12 +26,12 @@ describe('when calling org-formation perform tasks', () => {
         spawnProcessMock.mockReset();
         await PerformTasksCommand.Perform({...command, tasksFile: basePathForScenario + '1-deploy-cdk-workload-1target.yml' });
         spawnProcessAfterDeploy = spawnProcessMock.mock;
-        stateAfterDeploy = await s3client.getObject({Bucket: stateBucketName, Key: command.stateObject}).promise();
+        stateAfterDeploy = await s3client.send(new GetObjectCommand({Bucket: stateBucketName, Key: command.stateObject}));
 
         spawnProcessMock.mockReset();
         await RemoveCommand.Perform({...command, type: 'cdk', name: 'CdkWorkload' });
         spawnProcessAfterCleanup = spawnProcessMock.mock;
-        stateAfterCleanup = await s3client.getObject({Bucket: stateBucketName, Key: command.stateObject}).promise();
+        stateAfterCleanup = await s3client.send(new GetObjectCommand({Bucket: stateBucketName, Key: command.stateObject}));
     });
 
     test('1 workload was deployed', () => {
@@ -40,8 +39,8 @@ describe('when calling org-formation perform tasks', () => {
         expect(spawnProcessAfterDeploy.calls[0][0]).toEqual(expect.stringContaining('npx cdk deploy'));
     });
 
-    test('after deploy 1 region is found in state', () => {
-        const stateAsString = stateAfterDeploy.Body.toString();
+    test('after deploy 1 region is found in state', async () => {
+        const stateAsString = await stateAfterDeploy.Body.transformToString('utf-8');
         const state = JSON.parse(stateAsString);
         expect(state).toBeDefined();
         expect(state.targets).toBeDefined();