+++ test/integration-tests/tools/test-cfn-exports-backoff.ts
@@ -1,5 +1,5 @@
-import { CloudFormation } from "aws-sdk";
-import { AwsUtil } from "~util/aws-util";
+import { CloudFormationClient } from "@aws-sdk/client-cloudformation";
+import { AwsUtil, BoundCloudFormationClient } from "~util/aws-util";
 
 const stressExportsParallel = async () => {
   await Promise.all([stressExports(), stressExports(), stressExports(), stressExports(), stressExports(), stressExports()]);
@@ -22,5 +22,5 @@ const stressExports = async () => {
   await AwsUtil.GetCloudFormationExport("xxxxxxxx", "ccccccc", "eu-central-1", "");
 }
 
-AwsUtil.GetCloudFormation = () => { return Promise.resolve(new CloudFormation({ region: "eu-central-1" })); }
+AwsUtil.GetCloudFormationService = () => { return new BoundCloudFormationClient({ region: "eu-central-1" }, {}); }
 stressExportsParallel().then(x => console.log("done"));